Remove abstract key-val fsp and make indexedDB fsp operate on Uint8Arrs (#107590)
* Remove abstract key-val fsp and make indexedDB fsp operate on Uint8Arrs

* dont export
